+++ packages/react/components/input/package.json
@@ -27,7 +27,6 @@
     "test:watch": "vitest --watch"
   },
   "dependencies": {
-    "@nugudi/react-hooks-input": "workspace:*",
     "@nugudi/themes": "workspace:*",
     "@vanilla-extract/css": "^1.17.4",
     "@vanilla-extract/dynamic": "^2.1.5",
