+++ packages/react/components/input/src/index.ts
@@ -1,2 +1,6 @@
+// Components
 export { Input } from "./Input";
-export type { InputProps } from "./types";
+// Types
+export type { InputProps, UseInputProps, UseInputReturn } from "./types";
+// Hooks
+export { useInput } from "./useInput";
