+++ packages/react/components/input/src/types.ts
@@ -1,11 +1,25 @@
+import type * as React from "react";
+import type { InputHTMLAttributes } from "react";
+
+export interface UseInputProps extends InputHTMLAttributes<HTMLInputElement> {
+  invalid?: boolean; // aria-invalid용
+  errorMessage?: string;
+}
+
+export interface UseInputReturn {
+  inputProps: InputHTMLAttributes<HTMLInputElement> & {
+    "aria-invalid"?: boolean;
+    "aria-describedby"?: string;
+    "aria-required"?: boolean;
+  };
+  isFocused: boolean;
+}
+
 export type InputProps = {
   size?: "full";
   variant?: "default" | "filled";
   label?: string;
   isError?: boolean;
   errorMessage?: string;
-  // useInput props
-  isRequired?: boolean;
-  isReadOnly?: boolean;
-  isInvalid?: boolean;
+  invalid?: boolean; // aria-invalid용
 } & React.InputHTMLAttributes<HTMLInputElement>;
