--- packages/react/hooks/input/src/index.ts
@@ -1,5 +1,0 @@
-// Types
-export type { UseInputProps, UseInputReturn } from "./types";
-
-// Hooks
-export { useInput } from "./useInput";
