--- packages/react/hooks/input/vite.config.mts
@@ -1,33 +1,0 @@
-import { resolve } from "node:path";
-import { defineConfig } from "vite";
-import dts from "vite-plugin-dts";
-
-export default defineConfig({
-  plugins: [
-    dts({
-      include: ["src"],
-      outDir: "dist",
-      entryRoot: "src",
-      insertTypesEntry: true,
-      rollupTypes: false,
-      tsconfigPath: "./tsconfig.json",
-    }),
-  ],
-  build: {
-    lib: {
-      entry: resolve("src/index.ts"),
-      fileName: "index",
-      formats: ["es"],
-    },
-    rollupOptions: {
-      external: [
-        "react",
-        "react-dom",
-        "react/jsx-runtime",
-      ],
-    },
-    sourcemap: true,
-    minify: false,
-    outDir: "dist",
-  },
-});
